Update AutoPlay.js
Hi there Matt, I'm a fan of the Kissanime auto-play script and am fooling around with java script, I don't know how to code really good. I am just getting started so if you can improve on my ideas go ahead. 

I noticed that when you exit full-screen and you already auto-played a couple of times that the 'next' and 'previous' buttons would go to the corresponding pages based off of the last page you did not autoplay, or the last page that your browser refreshes into. So the way i tackled that was to use the SelectedIndex since it is changing corresponding to the current video. 

I also noticed that KissCartoon was encoding their video source differently than Kissanime making your script not work for KissCartoon, But I found that they use a function to decode it, so i used it as well.

I was trying to have the right and left keys do the same as the buttons on the website and switch pages based off the current selected index but was unable too.